Remove Cajita/Valija flag as an input to all cajoling services http://codereview.appspot.com/136053
Remove Cajita/Valija flag as an input to all cajoling services. Rely
only on the source itself to tell us whether it is Valija (the
default) or Cajita (via a "use cajita" pragma). Cajole all files to a
consistent output filename (*.out.js).

These changes allow the (Cajita-level) module loader to predictably
load a module regardless of whether it is Valija or Cajita. Because we
build our Valija module loader on top of Cajita, the Cajita loader did
not have the information of whether a module to be loaded is Cajita or
Valija; as far as it was concerned, everything is Cajita. Which is as
it should be -- Valija is just Cajita with a funky calling convention.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3846 360991c0-9f3b-0410-ba87-8b362a00edf4