bad html attributes downgraded to a warning http://codereview.appspot.com/164070
Prior to r3844, html with an invalid attr like
  <div title='bob's foo'>
would have the bad "foo'" attribute deleted with a warning.

r3844 changed that to be a fatal error.

Bad attributes are unfortunately common, for many reasons.
Usually because they're typos of some sort.

This CL downgrades the error to a warning.

R=mikesamuel@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3900 360991c0-9f3b-0410-ba87-8b362a00edf4