Issue 1180: Fix URI resolution and extract javascript URLs. http://codereview.appspot.com/180091
Prior, we resolved URIs in HTML inconsistently.
We completely ignored <base href> URIs, and instead got the URI
from the FilePosition.  This is problematic for content parsed
without debugging on.  This CL does not entirely address that problem
but adds a pass so that when the base URI is known, we correctly
apply it.

This change also consolidates code that finds and parses embedded
content: styles and scripts in HTML.

A following CL will have to thread the base URI reliably through from
our external APIs.  That will require coordination with clients, so is
out of scope for this change.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3931 360991c0-9f3b-0410-ba87-8b362a00edf4