Issue 1190: JSON.parse failing on Chrome http://codereview.appspot.com/224092
Chrome provides native JSON support.  This interacts badly with
revivers and replacers.  I think the solution is not to use revivers,
but to instead sanitize the input and output string.

This is the minimal fix.  It's ugly.  The problem that we're running
into is that methods which memoize state on checking canSetPub cause
infinite recursion.

R=erights@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4005 360991c0-9f3b-0410-ba87-8b362a00edf4