Refactor object constructors http://codereview.appspot.com/1033042
This changes the structure of ObjectConstructors to make it easier to
support setters/getters, e.g. ({ foo: bar, get baz() { } }).

This still needs some work to fix ancillary tests, and to possibly
refactor object constructor rewriting to have a rule for
object properties instead of having a bunch of loops.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4066 360991c0-9f3b-0410-ba87-8b362a00edf4