Make demos/container/container-dynamic.html work, by using cajita-module.js. http://codereview.appspot.com/1730041
container-dynamic.html as it stands does not work, because it does not serialize the new-module handler and <script> loading actions. It is changed to use cajita-module.js to manage module loading.

The set of JavaScript files served by the playground (and thus caja.appspot.com) and the set of files expected by container-dynamic.html are changed to match.

(Note that this will not actually work until this change is committed and caja.appspot.com is updated.)

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4141 360991c0-9f3b-0410-ba87-8b362a00edf4