Add arrows to clarify effect of buttons on text fields in Playground Source view. http://codereview.appspot.com/1744043
Add arrows to the labels of the "Load" and "Cajole" buttons in the Playground Source tab pointing, respectively, from the URL field to the source display and from the source display to off-to-the-right, to indicate what the actual data-flow / effect of clicking these buttons is.

Also remove 100% width declarations from these buttons, which eliminates layout problems on Firefox (newer versions) which are exacerbated by adding the arrows.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4154 360991c0-9f3b-0410-ba87-8b362a00edf4