Add support for maker idiom in ses.js http://codereview.appspot.com/1920043
The maker idiom often has the expression
  return freeze({prop: value, ...});
but none of the properties are whitelisted, making such objects
unusable by ES5/3 code.  This change introduces a method for use
only in the internals of ses.js that whitelists public own properties
of an object before freezing it.

Since ES5/3 code that uses that idiom gets translated into calls
to initializeMap, there's no need to change the idiom.

R=erights@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4216 360991c0-9f3b-0410-ba87-8b362a00edf4