Add ability to set URI policy to HostTools, and a vaguely-closer-to-the-right-thing default policy. http://codereview.appspot.com/1966044
Added method hostTools.setURIPolicy(p) where p is the uriCallback for Domita. Changed the default policy to, rather than return null, pipe everything through the cajoler (in a way which is about equally broken as the cajoler's own default URI policy).

The added tests also examine the URI rewriting done by the cajoler (since eventually the same policy will control both); this revealed that BrowserTestCase was not setting up the cajoling service's knowledge of its own URL, and so that has been fixed as well. I also put a warning on the zero-argument constructor of CajolingServlet.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4242 360991c0-9f3b-0410-ba87-8b362a00edf4