Speed up ES5/3 http://codereview.appspot.com/2203042
Sacrifices a bit more ES5/strict conformance in exchange for better
speed on ES3R browsers. The "___.wrap(function(dis___, ...) {"
translation correctly emulated the no-this-coercion rule from
ES5/strict, but at the price of at least 2 real function calls per
emulated function call. Instead, we now translate to
"___.f(function(...) {" with the original params, where ___.f is like
___.markFunc but also marks the object as safely callable directly as
a function from untrusted cajoled code. The operational difference is
that ___.f also sets "fn.i___ = fn;". The body of the function uses
the InnocentCodeRewriter trick: If the thisBinding on entry is the
global object, it binds dis___ to void 0. Else to this.

The new gotcha introduced by these optimizations is that primitive
this bindings are still implicitly wrapper as they should be for
non-strict code but which should be suppressed for strict code.

Moved the whitelisting of 'prototype', 'length', and 'name' on
individual functions into markFunc (which both ___.f and
markFreezeFunc call), though guarded with a kludgy check to avoid
circular initialization order problems.

Some inlining along the common path of DefineOwnProperty___.

"sharedImports" was accidentally global. Fixed.

For internal calls to Array.slice where it's clear the first arg
cannot be null or undefined, turned them into calls to slice.call.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4277 360991c0-9f3b-0410-ba87-8b362a00edf4