Add @Nonnull and @Nullable annotations to the lexer package. http://codereview.appspot.com/2864041
To be able to use findbugs to weed out problems in the HTML parser, we
need to tighten down its dependencies.  The lexer package is one of
these.

I added a package declaration for that package with a
@ParametersAreNonnullByDefault annotation, then ran findbugs and
whacked the null problems in that package.

This change also includes a bit of opportunistic cleanup of other
findbugs problems.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4328 360991c0-9f3b-0410-ba87-8b362a00edf4