Unicode escapes in identifiers bypass keyword checks and render as malformed programs after cajoling http://codereview.appspot.com/4178049
Words which are potential identifiers are checked against the Keywords list.  If a unicode escape 
occurs in the word, this keyword test fails.  However, the unicode escape is subsequently replaced
with the character, turning the word into an identifier.  Eg:

function wit\u0068 (a) {} is cajoled as function with(a) {}

This is not a security breach since the resulting code is not runnable, however, the resulting code
is not syntacically correct which violates one of Caja's invariants.

R=msamuel@google.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4369 360991c0-9f3b-0410-ba87-8b362a00edf4