neuter appengine http proxy cache http://codereview.appspot.com/4530056
appengine has an http proxy cache that makes it annoying to use
the caja playground there.  if you modify the contents at a url,
it can take several minutes before the change is visible to the
playground, and which version the playground gets can vary depending
on which instance of the proxy cache happens to answer the request.

this change reduces the annoyance by sending a Cache-Control header
with the playground's url fetches, which is honored by the proxy cache.

I tested this by manually uploading before/after versions to a
test app at appengine, and manually fiddling with them, verifying
that the "before" version notices changes slowly and inconsistently,
but the "after" version notices changes quickly and consistently.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4495 360991c0-9f3b-0410-ba87-8b362a00edf4