Adding support for handling IE conditional comments correctly <unassigned>
<![endif]> is currently treated as a DIRECTIVE and hence ignored by Caja parser
during the render phase. Fixing this by introducing the idea of an IE comment in
several parts of the code. The IE_COMMENT token captures everything starting
from <!--[if ....]> upto <![endif]-->, and is treated in exactly the same way as
the COMMENT token.

This patch also contains certain fixes to the NodesTest.java tests (related to
comments), by ensuring that the DomParser, HtmlLexer objects are created with
the right asXml parameter instead of using true, even in cases where we were
going to use the HTML rendering mode.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4531 360991c0-9f3b-0410-ba87-8b362a00edf4