Fix bug around writability of inherited properties. http://codereview.appspot.com/4625075
isWritable() was missing a test for the case when the x property is
marked writable but not yet fastpathed and an object o inherits x.
It was returning false when it should have returned true.

Also forced the result of an assignment to equal the rhs, not
the result of the setter.

R=erights@google.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4537 360991c0-9f3b-0410-ba87-8b362a00edf4