Port Domita to ECMAScript 5 as "Domado". http://codereview.appspot.com/4535118
Domado is a straight port of Domita, but written in pure ECMAScript 5 code; it is run cajoled under ES5/3. It is designed to also be capable of running in a pure SES environment, using strictly ES5 features to defend itself.

In order to support being cajoled and not having privileged access, Domado does not use any ___ properties internally (replacing them with WeakMaps) or externally (replacing them with separate objects providing privileged access). HtmlEmitter and Bridal have necessarily been ported in the same way.

R=mikesamuel@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4576 360991c0-9f3b-0410-ba87-8b362a00edf4