Make test suite assertGuestHtmlCorrect more robust. http://codereview.appspot.com/4839051
If HTML processing is broken such that the content of input <script> elements appears literally in the virtual document, then es53-test-cajajs-invocation will still pass, because it looks only for substrings which occur in the source.

This change breaks up the JS string literals so that the substring search will not match them.

Also added an assert I forgot for the id-suffix-matchup test I added while working on Domado.

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4582 360991c0-9f3b-0410-ba87-8b362a00edf4