Flash fixes http://codereview.appspot.com/4899044
* Replace Playground Flash demo with one that exercises the Flash bridge.
* Support disabling Flash in caja.js api.
* Make FlashBridgeClient always complain when it fails.
* Implement onNavigateToURL bridge.
* Remove caja/flash/Makefile.
* Add swfobject to third_party.

So, the weird thing here is the Flash bridge requires the swf files to
be on a host different from the container host.  For the Playground demo,
I chose the Caja repo on code.google.com as the other host, which means
the demo uses swf files from an earlier revision, and updating the demo's
swf files requires two commits.

R=jasvir@gmail.com, metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4598 360991c0-9f3b-0410-ba87-8b362a00edf4