remove useless try/catch in caja.js http://codereview.appspot.com/5323069
cajoledRunner in caja.js has a try/catch that does nothing but
log the exception and re-throw the exception, which doesn't get
caught.  this hurts debuggability because js debuggers show an
error backtrace for the re-throw statement, rather than a backtrace
for the original cause of the error, and it's difficult to get
to the original error's backtrace.

this change removes the useless and annoying try/catch.

R=ihab.awad@gmail.com, jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4655 360991c0-9f3b-0410-ba87-8b362a00edf4