Copy of Ka-Ping's html sanitizer patch. http://codereview.appspot.com/5305081
I LGTMed this at http://codereview.appspot.com/4641064/

Issue 4641064: Factor out a replaceable tagPolicy to enable customization of the sanitizer.

This exposes a hook named 'tagPolicy' in the sanitizer so that the
sanitizer can be customized to accept or transform particular tags,
for example, to accept <iframe> tags with a specific SRC, or transform
<embed> tags into <iframe> tags.
 
makeTagPolicy, sanitizeAttribs, and sanitizeWithPolicy are factored 
out so that they can be easily reused in custom tagPolicy functions.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4658 360991c0-9f3b-0410-ba87-8b362a00edf4