add --benchmark option to PluginCompilerMain http://codereview.appspot.com/5370041
This adds an option [--benchmark n] or [-bm n] to PluginCompilerMain,
making it easier to do exploratory timings from the commandline with
the bin/cajole script.

Previously there was a separate entry point PluginCompilerMain.Repeat,
but this is more convenient to use.

This also adds some more url policy options.

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4667 360991c0-9f3b-0410-ba87-8b362a00edf4