catch jobcache storage errors http://codereview.appspot.com/5491056
So it turns out it's actually possible to cajole jquery on the appspot
playground, without using the precajoler cache.  When the playground app
is warmed up, it only takes about 40s to cajole jquery.js, well within
the 60s deadline.

However, the cajoler throws a MemcacheServiceError when it tries to
store the result in the JobCache.  appengine limits memcache values
to 1M, and the JobCache entry for jquery.js is about 15M.

This change catches and logs those errors, so the cajoler will still
return a result rather than error 500.

This change also has a few changes to make it easier to test appengine
deployments.  In particular, "ant alt-playground" and "ant alt-deploy"
will use ../alt-appengine-web.xml instead of ./appengine-web.xml

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4690 360991c0-9f3b-0410-ba87-8b362a00edf4