speed up PipelineMaker http://codereview.appspot.com/5521055
Computing the cajoler's compilation pipeline is kind of slow,
takes about 500ms on my macbook.

Once a pipeline is computed, the result is cached, so a warmed-up
cajoler doesn't have this cost, but the first-run cost is still
kind of obnoxious.

This change simplifies the definition of tools in the compilation
pipeline, which reduces the size of the compilation state space that
needs to be searched to compute the pipeline.

With this change, computing the pipeline goes from 500ms to 25ms.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4766 360991c0-9f3b-0410-ba87-8b362a00edf4