create a test-index.html page http://codereview.appspot.com/5720052
There used to be only one browser test file, but now there
are a couple dozen tests meant to be run in browsers.

Running 'ant brserve' will start up a webserver that lets you
try most of the tests from any browser, but then it's pretty
inconvenient to enter the dozens of different test URLs into
a browser.

This change adds a simple test-index page with all the test URLs.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4800 360991c0-9f3b-0410-ba87-8b362a00edf4