fix irregularities in client-side uri policy http://codereview.appspot.com/5885044
Originally, the URI policy was the sole arbiter of what URIs are allowed,
but it turns out that people are not very good at writing correct URI
policies.  In particular, a naive "allow all URIs" is totally unsafe
because that also allows javascript: URIs.

So at some point we decided that regardless of what the URI policy said,
we'd also block javascript: URIs.  Basically, we first match URIs against
a safe whitelist of allowed schemes, and then the URI policy has a chance
to further rewrite or reject the URI.

So.. that was implemented, but not comprehensively.  Some places applied
the URI policy with pre-filtering, some applied it without.

This change fixes the irregularities in client-side URI rewriting so that
all URIs get pre-filtered before applying the URI policy.

This doesn't fix handling of server-side URI policy, but that isn't really
used now, and the change has some complications, so I'm deferring it.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4819 360991c0-9f3b-0410-ba87-8b362a00edf4