escape <!-- in html-sanitizer http://codereview.appspot.com/5985057
html-sanitizer.js has some character sequences like '<!--' and '</script>'
that cause problems for some script inlining processes.  this is arguably
a bug with the inlining process, but most scripts do not cause problems,
and it's often harder to fix the inliner than it is to fix the problem
scripts.

so, this makes html-sanitizer.js friendlier to script inlining by
making sure that '<!' and '</' never appear literally.

(though of course this doesn't solve all potential inlining problems.)

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4848 360991c0-9f3b-0410-ba87-8b362a00edf4