misc changes to browser test framework http://codereview.appspot.com/6161046
- build.xml has logic to remove generated files that have version
in them, so that incremental builds work correctly if your svn
version changes.  but it wasn't removing all generated files.
this patch fixes that.

- GeneralBrowserTest takes several minutes to run, and it isn't easy
to see what parts are slow.  this patch adds a print before
each browser test page, to make it easier to track test speed.

- in cases where we run both es5 and es53 browser tests, we were
running the es53 version first.  this patch flips it so that es5
runs first, because es5 tends to be faster, and a test that fails
for both es5 and es53 will fail more quickly.

- GeneralBrowserTest has a number of tests that fiddle with version
strings in the caja js files, so they do a closeWebDriver()
with the comment "Need a browser with an empty cache".  This appears
to be not true.  Removing the closeWebDriver() statements does
not seem to affect the validity of the tests.

- es53-test-cajajs-never-starts.js relies on a 2-second timeout to
detect caja.js startup failure (since it's difficult to trap all the
errors that might be thrown).  1 second seems to be enough.

these last two items reduce runtime of GeneralBrowserTest on my
laptop from 3m20s to 3m00.

R=ihab.awad@gmail.com, metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4869 360991c0-9f3b-0410-ba87-8b362a00edf4