speed up junit http://codereview.appspot.com/6250057
We're running junit with fork="yes" because we want the junit vm to
have different properties than the ant vm.  However, we're spawning
a new vm for each junit test, which is unnecessary.

This adds forkmode="once" to the junit tests.

This reduces 'ant runtests' on my laptop from about 11 minutes to
about 7 minutes.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4889 360991c0-9f3b-0410-ba87-8b362a00edf4