fix ancillary test failures http://codereview.appspot.com/6280046
runtests still passes if failing tests are marked @FailureIsAnOption,
but it also passes if tests fail in com.google.caja.ancillary.
Currently, there are a few trivial failures in ancillary.

I don't see much value in ignoring ancillary test failures, since
it's pretty easy to manually mark tests with @FailureIsAnOption,
so I'm removing the special case for ancillary.

The test failures are:

- Some tests expect that catch(e) masking an outer e would be an
error, but we downgraded that to a warning.

- One test expected a['foo'] to be optimized to a.foo, but we
removed that optimization for Closure compatibility.

- One test expects to see "i++" instead of "++i", in a context where
it doesn't matter which gets used.

- There were tests for the jsdoc servlet, which was removed, but
the tests were not removed.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4897 360991c0-9f3b-0410-ba87-8b362a00edf4