make some browser tests dom-less http://codereview.appspot.com/6305052
this is a minor speedup to browser tests.

attachDocument in es53 mode takes about 2s, which we can
avoid by running a test in a dom-less guest.

I started doing this thinking I'd find more places where we
can run dom-less, but I found less than I hoped for,
so this is a pretty insignificant speedup, but I've already
done the work so I figure I might as well commit it.

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4906 360991c0-9f3b-0410-ba87-8b362a00edf4