faster ParseTreeNode visitor http://codereview.appspot.com/6305076
ParseTreeNode has tree-walk methods called acceptPreOrder and
acceptPostOrder that are worst-case O(n) space and O(n**2) time.

The O(n) space is because ParseTreeNode does not have a link
to the node's parent, but some visitors want to know a node's
parent, so the tree-walker constructs a new AncestorChain object
to hold a node and its parent, then gives that AncestorChain to
the visitor.  This is wasted effort when a visitor doesn't care
about a node's parent.

The O(n**2) time is because the tree-walker tries to be robust
against mutations to the tree that affect the current node's
place in the tree, so before (or after) visiting each node,
it checks whether the node is still a child of the parent,
which requires a linear search of the parent's children list.

So, previously I had replaced one instance of acceptPreOrder
with a faster visitPreOrder method.

This patch expands on that, makes visitPreOrder faster and
simpler, and also uses it in more places.

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4914 360991c0-9f3b-0410-ba87-8b362a00edf4