fix some servlet issues http://codereview.appspot.com/6304091
1. 'ant clean runserver' fails because we're no longer building the
tools servlet.  in the appengine logs, there are zero uses of the
tools in the past 90 days, so I'm just deleting them from web.xml.
If anyone complains, I can re-enable it and fix the build.

2. In some cases, IE will disbelieve content-type: application/json
and interpret the result as html instead.  This is not particularly
exploitable, since caja.appspot.com has no cookies and no authority,
but if someone deploys the caja servlet to their own host, they
might be vulnerable to the XSS.
- to fix that for IE >= 8, I'm emitting X-Content-Type-Options: nosniff
- to fix that for IE7, I'm rejecting request URLs that contain ';'
- IE6 users lose, they're vulnerable.  Fixing that is complicated
and doesn't seem worth it.

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4928 360991c0-9f3b-0410-ba87-8b362a00edf4