fix xerces linkage issue http://codereview.appspot.com/6361043
Currently we ask the JRE for any DOM implementation, and we have a
tendency to get the JRE's copy of an old version of xerces, which has
trouble with some of the weird things we do with DOM, such as
  createDocument(null, null, doctype)

and our workaround was to specify CLASSPATH=.../xercesImpl.jar when
running ant, which is kind of fragile.

This patch instead specifically uses the right xerces DOM
implementation.

It might be possible to fix our DOM usage so that it doesn't blow
up on other DOM implementations, and I started doing that, but
fixing the createDocument() call exposed other fatal DOM errors,
and so it seems simpler to just link with a known DOM impl.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4957 360991c0-9f3b-0410-ba87-8b362a00edf4