force autocomplete=off harder http://codereview.appspot.com/6432044
It's possible to create forms and input elements with autocomplete=on
by setting the attribute in js.  This change neuters that.

I've marked a weird TODO in the browser test.  For some reason, in firefox
with es5, form.setAttribute() fails because form.setAttribute is not a
function.  It works in chrome.  That failure seems to be irrelevant to the
autocomplete problem, so I'll try to fix that in another CL.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4967 360991c0-9f3b-0410-ba87-8b362a00edf4