fix Closure Compiler breakage http://codereview.appspot.com/6432052
Closure Compiler has an option to supply a class to enforce particular
coding/naming conventions.  We comply with the default CodingConvention,
but we're violating GoogleCodingConvention.

This change fixes that violation, and also makes our build use
GoogleCodingConvention so we can catch other errors early.

Also, fixed a potential false positive build success.  We
run tests in children in parallel, and then check for success
by checking their test outputs, but it's possible for a test
run to fail without generating any test outputs.  This
change makes the test run fail if any of the children fail.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4969 360991c0-9f3b-0410-ba87-8b362a00edf4