add failing testcase for js parser bug http://codereview.appspot.com/6428071
This is for
http://code.google.com/p/google-caja/issues/detail?id=1423

The js parser parses {{} /x/} incorrectly, because the js lexer
tries to decide whether / is division or regexp, and that decision
can't be made reliably in the lexer.

We need to move that decision to the parser, but that's hard
because the js parser is connected to the js lexer via an awkward
TokenQueue abstraction.  I think TokenQueue needs to be ripped
out here, which is a nontrivial amount of work.

This is a harmless bug, because
- wrong parse does not affect security
- wrong parse is unlikely in normal code
- it's easy to rewrite js to force the right parse

So I'm adding a failing testcase for the bug, and I'm going
to close the bug as wontfix.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4970 360991c0-9f3b-0410-ba87-8b362a00edf4