fix build.xml svn version check http://codereview.appspot.com/6442066
If an edit changes your svn version from 9999 to 9999m, build.xml
has a rule that will delete all the old js files with the old
version number baked in.

"all" is a lie.  In particular, browser test files are not deleted,
so browser tests fail with a complaint about version skew.

It used to work, but it broke when test files were moved out of
ant-lib into ant-testlib.

This change fixes that, by cleaning out ant-testlib too.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@4984 360991c0-9f3b-0410-ba87-8b362a00edf4