Implement document.title correctly per HTML5. http://codereview.appspot.com/6459089
The previous implementation returned the <title> node rather than the title
text, which is entirely wrong as far as I know, and also causes an incidental
error in the jQuery test suite. This implementation is according to the HTML5
spec, except insofar as our pseudo-DOM (to be replaced soon) makes that
difficult.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5012 360991c0-9f3b-0410-ba87-8b362a00edf4