upgrade to appengine-java-sdk-1.7.1 http://codereview.appspot.com/6501136
the one interesting thing in this change is appengine-web.xml now
requires you to have a <threadsafe> declaration.  the default has
been false, so I set the explicit value to false.  we might be
able to flip it to true, but that requires a nontrivial amount of
code auditing, which I'm deferring.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5053 360991c0-9f3b-0410-ba87-8b362a00edf4