relax number of css font-names http://codereview.appspot.com/6544050
our server-side css pattern matcher limits [foo*] patterns to 6
repetitions, which means a font-family: spec with more than 6 fonts
will fail to parse.

so this change removes the limit.  I think the matcher is O(n**2)
space and time, where n is the number of input tokens, but n will
typically be small.  it's unlikely that real uses of css will
have a large enough n to be noticeable.  and this is not the only
way that someone can craft pathologically expensive input, so I
don't think it's really worth fixing this.

this fixes https://code.google.com/p/google-caja/issues/detail?id=1540

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5061 360991c0-9f3b-0410-ba87-8b362a00edf4