Implement <body onload="..."> for client-side. http://codereview.appspot.com/6499129
<body onload="..."> is a special case because, while most HTML onfoo
attributes correspond to a feral node's .onfoo property, those on <body>
correspond to the tame window's .onfoo property. Since the tame window
is entirely our own invention, the handler-attribute-to-handler-property
behavior must be (and is now) implemented separately.

Since the assignment to window.onload is a side effect of Domado's
setAttribute, HtmlEmitter must also be modified so that it uses
Domado's setAttribute rather than sanitizing and passing to the host
setAttribute.

Provision is made for extending this support to event handlers other
than onload, but the rest of Domado does not yet support any such.

Fixes Issue 1478
<http://code.google.com/p/google-caja/issues/detail?id=1478>

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5063 360991c0-9f3b-0410-ba87-8b362a00edf4