Autoswitch between ES5 and ES53 http://codereview.appspot.com/6495136
Actually autoswitch between ES5 and ES53 depending on a flag and what the browser supports.

If es5Mode = undefined, try SES and if the browser does not support maxAcceptableSeverity failover to ES53.
If es5Mode = true, try SES and if the browser does not support maxAcceptableSeverity fail and don't run.
If es5Mode = false, just run es53.

The forceES5Mode flag is deprecated except for testing.  If it is set, value is passed to
es5Mode and the maxAcceptableSeverity is set to SAFE_SPEC_VIOLATION.

R=felix8a@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5094 360991c0-9f3b-0410-ba87-8b362a00edf4