reading from /Users/felixz/caja/trunk/.appspot-change.noembed
remove noembed support
http://codereview.appspot.com/6737049

when I implemented <embed> support, I also implemented <noembed>
support, and now I kinda regret it, because there are hazards
around any rawtext element.  none of the modern/supported browsers
will have problems, but there are potential problems if someone
attempts html rewriting without awareness of the rawtext problems,
so I think it's safer to just disallow noembed.

I don't know of any real demand for noembed, so this shouldn't
be a problem for users.  Some cut+paste html snippets have both
embed and noembed forms; we'll delete the noembed and its contents,
which is the right thing to do.

so this CL blacklists noembed and removes support for it.

R=kpreid.switchb.org@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5103 360991c0-9f3b-0410-ba87-8b362a00edf4