embeds can throw an unexpected error http://codereview.appspot.com/6734058
if a container doesn't support Flash and a guest tries to use <embed>
or <object>, we throw an error during html-emitter calling handleEmbed,
which isn't caught, which breaks the guest completely rather than
degrading gracefully.

this CL makes handleEmbed continue without error, which means the
alternate content inside an <object> will remain displayed, and the
guest will continue to run.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5105 360991c0-9f3b-0410-ba87-8b362a00edf4