Enable es53-test-client-uri-rewriting.js in ES5 mode. http://codereview.appspot.com/6813066
es53-test-client-uri-rewriting.js was disabled for an obsolete reason
and because it depended on intercepting frame.imports.emitCss___, which
is not used in SES mode.

Additionally, install jsunitCallback and function names in order to
make it easier to determine the source of console output. jsunitCallback
needed a small change to be usable from a host-page context.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@5128 360991c0-9f3b-0410-ba87-8b362a00edf4